Create abstract SlangSensor
